Improved game object system, level generation with renderable objects is done!

+++ core/src/com/tuguzteam/netdungeons/input/RotationZoomGestureListener.kt
@@ -70,7 +70,10 @@ class RotationZoomGestureListener(private val gameScreen: GameScreen) : KtxGestu
             val progress = if (rotationTime < 0) 1f else 1f - rotationTime / ROTATION_DURATION
             val angle = Interpolation.fade.apply(0f, ROTATION_ANGLE, progress)
             camera.rotateAround(
-                vec3(x = gameScreen.playerPosition.x, z = gameScreen.playerPosition.y),
+                vec3(
+                    x = gameScreen.playerPosition.x.toFloat(),
+                    z = gameScreen.playerPosition.y.toFloat(),
+                ),
                 Vector3.Y,
                 (angle - rotationOldAngle) * sign,
             )
